+++ ninja-maven-plugin/src/main/java/ninja/NinjaRunMojo.java
@@ -27,52 +27,38 @@ public class NinjaRunMojo extends AbstractMojo {
     
     public final String CLASSES_DIRECTORY = File.separator + "target" + File.separator + "classes";
     
-    public final String NINJA_JETTY_CLASS_AND_CLASSPATH = "ninja.standalone.NinjaJetty";
-    
     /**
      * @parameter expression="${project}"
      * @required
      * @readonly
      */
     protected MavenProject mavenProject;
-    
-    /** 
-     * @parameter default-value="${plugin.artifacts}" 
-     */
-    private List<Artifact> pluginArtifacts;
 
     @Override
     public void execute() throws MojoExecutionException {
         
-        String completeClassesDir = new File("").getAbsolutePath() + CLASSES_DIRECTORY;
-
-        System.out
-                .println("------------------------------------------------------------");
-        
         List<String> classpathItems = Lists.newArrayList();
-        //classpathItems.add(".");
+        
+        
+        
+        String directoryWithCompiledClassesOfThisProject 
+                = System.getProperty("user.dir") + CLASSES_DIRECTORY;
+        
+        classpathItems.add(directoryWithCompiledClassesOfThisProject);
 
-        classpathItems.add(completeClassesDir);
-        
-        //Artifact artifact = Artifact
-        
-        for (org.apache.maven.artifact.Artifact artifact : pluginArtifacts) {
-            //classpathItems.add(artifact.getFile().toString());  
-        } 
-        
         for (org.apache.maven.artifact.Artifact artifact : mavenProject.getArtifacts()) {
             classpathItems.add(artifact.getFile().toString());           
         }       
         
-        Path path = FileSystems.getDefault().getPath(completeClassesDir);
+        Path path = FileSystems.getDefault().getPath(directoryWithCompiledClassesOfThisProject);
         
         try {
-            NWatchAndTerminate nWatchAndTerminate = new NWatchAndTerminate(
+            
+            WatchAndRestartNinjaMachine nWatchAndTerminate = new WatchAndRestartNinjaMachine(
                     path,
-                    true, 
-                    NINJA_JETTY_CLASS_AND_CLASSPATH, 
                     classpathItems);
             nWatchAndTerminate.processEvents();
+            
         } catch (IOException e) {
             // TODO Auto-generated catch block
             e.printStackTrace();